refactor: 서평 메모 폴더링 변경
