+++ lib/review/screen/book_search_screen.dart
@@ -4,7 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:typed/common/const/index.dart';
 import 'package:typed/common/index.dart';
 import 'package:typed/review/viewmodels/book_viewmodels.dart';
-import 'package:typed/review/model/book_model.dart';
+import 'package:typed/review/models/book_model.dart';
 
 class BookSearchScreen extends ConsumerStatefulWidget {
   const BookSearchScreen({super.key});
