+++ lib/review/screen/review_input_screen.dart
@@ -4,10 +4,10 @@ import 'package:go_router/go_router.dart';
 import 'package:typed/common/const/app_colors.dart';
 import 'package:typed/common/const/app_themes.dart';
 import 'package:typed/common/index.dart';
-import 'package:typed/review/model/book_model.dart';
-import 'package:typed/review/model/lock_enum.dart';
+import 'package:typed/review/models/book_model.dart';
+import 'package:typed/review/models/lock_enum.dart';
 import 'package:typed/review/viewmodels/book_viewmodels.dart';
-import 'package:typed/review/review_provider.dart';
+import 'package:typed/review/viewmodels/review_provider.dart';
 
 class ReviewInputScreen extends ConsumerStatefulWidget {
   const ReviewInputScreen({super.key});
