+++ lib/review/screen/review_list_screen.dart
@@ -4,9 +4,9 @@ import 'package:go_router/go_router.dart';
 import 'package:typed/common/const/index.dart';
 import 'package:typed/common/layout/default_layout.dart';
 import 'package:typed/common/widgets/app_bar/custom_app_bar.dart';
-import 'package:typed/review/model/lock_enum.dart';
-import 'package:typed/review/model/review_model.dart';
-import 'package:typed/review/review_provider.dart';
+import 'package:typed/review/models/lock_enum.dart';
+import 'package:typed/review/models/review_model.dart';
+import 'package:typed/review/viewmodels/review_provider.dart';
 
 final dummyReviewsProvider = Provider<List<Review>>((ref) {
   return [
